requirements.txt
PyJWT>=2.8.0
cryptography>=41.0.0
requests>=2.31.0
PyYAML>=6.0
lxml>=4.9.3
